Interview edits
Make it easier for future copy and pastes so that ctrl + / will uncomment the video section, and that "Video of Interview" does not appear if video is not available.   Also added No_vid text to say witness did not consent to video recording.   Also removed lecture video from Prof. Amethyst